Save and restore scroll positions
Save the scroll position for each page in a cookie when leaving the page,
and scroll to the stored position upon load.

There is a discrepancy between the position in show and edit modes
because it's hard to reliably find the corresponding positions when text
size and spacing changes. Perhaps a future improvement would be to count
paragraphs instead.

This introduces the third-party library js-cookie.